smartbuffer: fixed optimization
The optimization of the minibuffer (sBuffer_single)
had / has a big issue.
This issue also lead to the incredibly good results in the tests.
The issue was that the reallocation was not done properly.

This change fixes most of this,
but as each reallocation on a sBuffer_single
could lead to a change of address,
the sBuffer_single_ptr must be updated every time.

It might be wise to think about an implementation,
where it does not change it's own address,
because else this could lead to severe issues
regarding shared access to a sBuffer_single_ptr.