smartbuffer: revert optimization
I decided two switch back to two allocations.
I did NOT revert all changes and kept most of them.
SIMPLE_ARRAY is NOT used.

The reason for this change is that the single allocation
lead to safety concerns.
On each reallocation the address of the sBuffer_single_ptr
could change. Thus managing shared access has become dangerous.
Consider the following:

    sBuffer_single_ptr mydata = ...;
    sBuffer buf = ...;

    // add reference of mydata to buf
    sBuffer_add(buf, mydata);

    // change mydata - Note: address of mydata may change!!!
    mydata = sBuffer_single_add(mydata, ...);

If mydata changed its address, the address must also be updated
inside buf else buf has a bad pointer
-> may lead to segmentation fault!!

+ Added new testresults for this new status.
Of course the performance got worse,
but it was even worse when SIMPLE_ARRAY was used.

For performance one should mainly look at the difference between
- minibuffer / sBuffer_single
- smartbuffer
as the "bad" performance of the smartbuffer is mainly
because of the "bad" performance of the minibuffer / sBuffer_single.
The optimization try from before shows this clearly.
std::string on my macos and on my ubuntu has just one allocation,
but does not implement shared handling / usage count / reference count.
Furthermore the std::string on my ubuntu (glibc) has an exponential memory growth.