smartbuffer: test more iterations
Changed the test string_append_cont so that it can be run "normally".
The iterations from google benchmark do NOT play a big role anymore
and can be used to get better average times.

If the google benchmark iterations are too high,
the program will be killed by macos.
As I saw that the process uses a lot of RAM,
I presume that the excesive malloc calls
lead to the process getting killed,
because it uses too much memory.

It may help to add some malloc_trim() or so...

The tests are not that good anymore.
I also have a script that does graphs,
so now I have images!!