test on windows + fixed bad memory leak
Made a test on windows.
The CPU time does not look realiable,
so look at the real time, which seems to be correct.
The only strange thing is that the minibuffer takes so long.
Thus on the graph it seems that the other take 0 time,
which is not true - the scale is just bigger...

As this test used a lot of RAM (up to nearly 4 GB),
I looked for memory leaks and found a pretty bad one.
Thus I fixed it and made a test on MacOS.