No iteration limit anymore
As the memory leak is gone,
the memory used is way less
and the OS does not kill the process.
Thus we can do as many iterations as we want.

The library now handles the number of iterations.

I'M still questioning the result,
as Minibuffer should not be faster than std::string...
The only reason I could think of is that the Minibuffer
uses the C function realloc and std::string always
copies to a new allocation.