use min warmup time again + test
Used min warmup time again in order to stabilize results,
but they keep being strange.

The Minibuffer is very slow on windows and too fast on linux.
The first iteration of BM_SmartBuffer is too slow on Windows,
it simply cannot be that appending 2 x 4000 bytes is faster
than appending 1 x 4000 bytes (and with a factor ~4).
This only happened on Windows VM ARM64 though...
Not even on Windows VM x64...