fix windows build
There was a mistake with SinglePtr::New(0) and then realloc.
A bad (uninitialized) address was passed to realloc.
Apparantly only Windows could not handle this.
This caused some "quiet" killing of the process,
which resulted in that it seemed like only one test was run,
even tough the second one crashed (quietly)...

Fixed this and initialized the pointer with NULL.
Also managed to build with mingw and with VS Code Debug.