string_append_cont: different growing policy
The std::string was in net_heap_growth as good as the Smartbuffer
in the last tests.
The reason is simple:
When std::string uses an exponential growth policy
and we always append the same amount of data with the same factor (2),
as std::string, all memory that it reserves, will be used at the end.

In Short: The test grew exactly the same as the string,
so all memory was used in the end.

Now I changed the growth factor,
and loolit looks totally different now...